EM-3140: Moving R26 Test to a new Class.
